fix helper text on wallet flow
